bot/requirements.txt
pyTelegramBotAPI==4.15.3
python-dotenv==1.0.0